Last fixes

+++ Security Room/Assets/PersonInfo.cs
@@ -5,8 +5,8 @@ public class PersonInfo {
 	public String id;
 	public int age;
 	public string gender;
-	public int height;
-	public int weight;
+	public double height;
+	public double weight;
 
 	public static PersonInfo createFromJsonString(string jsonString) {
 		return JsonUtility.FromJson<PersonInfo> (jsonString);
